Fix Python 3 issues
* Add dependency to six in requirements.txt
* Replace long(500) with 500. long type was replaced with the int
  type in Python 3.
* Fix a commented print statement
* Replace xrange() with range() or six.moves.range() depending on the
  size of the range.
* Use universal_newlines to get stdout as text (Unicode) on Python 3
  instead of bytes. Use also communicate() which is more reliable and
  standard than stdout.read().
* Replace long(str) with int(str)

+++ hardware/diskinfo.py
@@ -27,7 +27,7 @@ def sizeingb(size):
 
 def disksize(name):
     size = open('/sys/block/' + name + '/size').read(-1)
-    return sizeingb(long(size))
+    return sizeingb(int(size))
 
 
 def disknames():
